Find Nash equilibria of a normal-form game.

solve_nfg_char() finds the pair of the best responses of a normal-form game defined with payoff functions as character strings by normal_form().

solve_nfg_char(
  game,
  plot = TRUE,
  mark_NE = FALSE,
  quietly = FALSE,
  color_palette = "Set1"
)

Arguments

game

A "normal_form" class object created by normal_form().

plot

A logical value to determine whether the figure of the best response correspondences will be displayed. Default is TRUE.

mark_NE

A logical value to control if the NE (if any) will be marked in the best response plot, which will be displayed (only displayed when plot = TRUE). Default is FALSE.

quietly

A logical value to determine if the equilibrium will be kept in the returned list without being printed on screen. Default is FALSE.

color_palette

A color palette to be used. Default is "Set1".

Value

A list containing the pair of the best responses of two players (NE) and the plot of best response correspondences.

See also

Author

Yoshio Kamijo and Yuki Yanai yanai.yuki@kochi-tech.ac.jp