Find subgame perfect equilibria of a sequential-form game

solve_seq() finds subgame perfect equilibrium outcomes of a sequential-form game. This is a wrapper function of solve_seq_matrix, solve_seq_char, and solve_seq_fcn.

solve_seq(
  game,
  show_table = TRUE,
  cell_width = NULL,
  mark_br = FALSE,
  cons1 = NULL,
  cons2 = NULL,
  cons_common = NULL,
  precision = 1L,
  quietly = FALSE
)

Arguments

game
A "sequential_form" class object created by seq_form().

show_table
A logical value. If TRUE, the table of the game will be displayed. Default is TRUE.

cell_width
A number specifying the cell width of the game matrix. The unit is pixel.

mark_br
A logical value. If TRUE, the follower's best response to each of the leader's strategy is marked. Default is FALSE.

cons1
A named list of parameters contained in game$payoff$p1 that should be treated as constants, if any.

cons2
A named list of parameters contained in game$payoff$p2 that should be treated as constants, if any.

cons_common
A named list of parameters contained in game$payoff$p1 and game$payoff$p2 that should be treated as constants, if any. If cons1 and cons2 are exactly same, you can specify cons_common instead of specifying both cons1 and cons2.

precision
A natural number specifying the precision of numerical approximation. The value n approximately means that the approximation is correct up to the Nth decimal place. The default value is 1.

quietly
A logical value that determines whether the equilibrium will be kept in the returned list without being printed on screen. Default is FALSE.

Value

A list containing subgame perfect equilibrium outcomes (if any), and the game table (if available).

Author

Yoshio Kamijo and Yuki Yanai yanai.yuki@kochi-tech.ac.jp